Move single Safe creation to test utils
