+++ scripts/utils/internals.js
@@ -1,7 +1,6 @@
 module.exports = function(web3 = web3, artifacts = artifacts) {
   const util = require("util")
   const lightwallet = require("eth-lightwallet")
-  const assert = require("assert")
 
   const GnosisSafe = artifacts.require("GnosisSafe.sol")
   const MultiSend = artifacts.require("MultiSend")
@@ -78,28 +77,6 @@ module.exports = function(web3 = web3, artifacts = artifacts) {
       ADDRESS_0,
       sigs
     )
-  }
-
-  const deploySafe = async function(gnosisSafeMasterCopy, proxyFactory, owners, threshold) {
-    const initData = gnosisSafeMasterCopy.contract.methods
-      .setup(owners, threshold, ADDRESS_0, "0x", ADDRESS_0, ADDRESS_0, 0, ADDRESS_0)
-      .encodeABI()
-    const transaction = await proxyFactory.createProxy(gnosisSafeMasterCopy.address, initData)
-    return getParamFromTxEvent(transaction, "ProxyCreation", "proxy", proxyFactory.address, GnosisSafe, null)
-  }
-
-  // Need some small adjustments to default implementation for web3js 1.x
-  async function getParamFromTxEvent(transaction, eventName, paramName, contract, contractFactory, subject) {
-    // assert.isObject(transaction)
-    if (subject != null) {
-      logGasUsage(subject, transaction)
-    }
-    let logs = transaction.logs
-    if (eventName != null) {
-      logs = logs.filter(l => l.event === eventName && l.address === contract)
-    }
-    assert.equal(logs.length, 1, "too many logs found!")
-    return logs[0].args[paramName]
   }
 
   async function createLightwallet() {
@@ -209,15 +186,9 @@ module.exports = function(web3 = web3, artifacts = artifacts) {
     return execTransaction
   }
 
-  function logGasUsage(subject, transactionOrReceipt) {
-    const receipt = transactionOrReceipt.receipt || transactionOrReceipt
-    console.log("    Gas costs for " + subject + ": " + receipt.gasUsed)
-  }
-
   return {
     waitForNSeconds,
     execTransaction,
-    deploySafe,
     encodeMultiSend,
     createLightwallet,
     signTransaction,
