+++ test/test-utils.js
@@ -1,6 +1,7 @@
 const TokenOWL = artifacts.require("TokenOWL")
 const TestToken = artifacts.require("DetailedMintableToken")
 const { toErc20Units } = require("../scripts/utils/printing_tools")
+const ADDRESS_0 = "0x0000000000000000000000000000000000000000"
 
 const prepareTokenRegistration = async function(account, exchange) {
   const owlToken = await TokenOWL.at(await exchange.feeToken())
@@ -22,8 +23,35 @@ const addCustomMintableTokenToExchange = async function(exchange, symbol, decima
     token: token,
   }
 }
+const deploySafe = async function(gnosisSafeMasterCopy, proxyFactory, owners, threshold) {
+  const initData = gnosisSafeMasterCopy.contract.methods
+    .setup(owners, threshold, ADDRESS_0, "0x", ADDRESS_0, ADDRESS_0, 0, ADDRESS_0)
+    .encodeABI()
+  const transaction = await proxyFactory.createProxy(gnosisSafeMasterCopy.address, initData)
+  return getParamFromTxEvent(transaction, "ProxyCreation", "proxy", proxyFactory.address, GnosisSafe, null)
+}
+
+// Need some small adjustments to default implementation for web3js 1.x
+async function getParamFromTxEvent(transaction, eventName, paramName, contract, contractFactory, subject) {
+  // assert.isObject(transaction)
+  if (subject != null) {
+    logGasUsage(subject, transaction)
+  }
+  let logs = transaction.logs
+  if (eventName != null) {
+    logs = logs.filter(l => l.event === eventName && l.address === contract)
+  }
+  assert.equal(logs.length, 1, "too many logs found!")
+  return logs[0].args[paramName]
+}
+
+function logGasUsage(subject, transactionOrReceipt) {
+  const receipt = transactionOrReceipt.receipt || transactionOrReceipt
+  console.log("    Gas costs for " + subject + ": " + receipt.gasUsed)
+}
 
 module.exports = {
   prepareTokenRegistration,
   addCustomMintableTokenToExchange,
+  deploySafe,
 }
